tvOS unit tests pass.

+++ Branch/BNCURLBlackList.m
@@ -158,7 +158,7 @@
         BNCLogDebugSDK(@"No new BlackList refresh found.");
     } else {
         BNCLogDebugSDK(@"BlackList refresh result. Error: %@ status: %ld body:\n%@.",
-            operation.error, operation.HTTPStatusCode, responseString);
+            operation.error, (long) operation.HTTPStatusCode, responseString);
     }
     if (operation.error || operation.responseData == nil || operation.HTTPStatusCode != 200) {
         self.error = operation.error;
